# user-app/src/main/resources/static/api/swagger.yml
openapi: 3.0.3
info:
  title: Scalable Task Manager User App
  version: "1.0"
paths:
  /v1/auth/register:
    post:
      tags:
        - auth
      summary: Registers a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserDTO"
      responses:
        200:
          description: "User was successfully registered"
          content:
            text/plain:
              schema:
                type: string
        400:
          description: "Client did not send the server info correctly"
        500:
          description: "A generic backend error occurred"
  /v1/auth/login:
    post:
      tags:
        - auth
      summary: Login a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserDTO"
      responses:
        200:
          description: "User was successfully registered"
          content:
            text/plain:
              schema:
                type: string
        400:
          description: "Client did not send the server info correctly"
        401:
          description: "User or password weren't correct"
        500:
          description: "A generic backend error occurred"
  /v1/auth/update:
    put:
      tags:
        - auth
      summary: Update existing user
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDTO"
      responses:
        200:
          description: "Password was successfully updated"
        400:
          description: "Client did not send the password correctly"
        401:
          description: "Authentication failed"
        500:
          description: "A generic backend error occurred"
  /v1/users/{username}:
    get:
      summary: Get existing user
      operationId: getUser
      tags:
        - get
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the user to get
      responses:
        200:
          description: "User was successfully retrieved"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        400:
          description: "Client did not send the server info correctly"
        500:
          description: "A generic backend error occurred"
components:
  schemas:
    RegisterUserDTO:
      type: object
      required:
        - fullName
        - username
        - password
      properties:
        fullName:
          type: string
        username:
          type: string
        password:
          type: string
    LoginUserDTO:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    UpdateUserDTO:
      required:
        - newPassword
      properties:
        newPassword:
          type: string
    UserDTO:
      type: object
      required:
        - id
        - fullName
        - username
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        username:
          type: string